specify separate project module
