+++ solo-leveling-ui-service/src/main/bundles/dev.bundle
