+++ solo-leveling-ui-service/src/main/frontend/index.html
