+++ solo-leveling-ui-service/src/main/frontend/styles/home-view.css
