+++ solo-leveling-ui-service/src/main/frontend/styles/swiped.css
