+++ solo-leveling-ui-service/src/main/frontend/themes/my-app/components/vaadin-text-area.css
