+++ solo-leveling-ui-service/src/main/frontend/themes/my-app/theme.json
