+++ solo-leveling-ui-service/src/main/frontend/themes/my-app/views/my-view-view.css
