+++ solo-leveling-ui-service/src/main/java/com/sleepkqq/sololeveling/ui/localization/LocalizationException.java
