+++ solo-leveling-ui-service/src/main/java/com/sleepkqq/sololeveling/ui/service/TgHashService.java
@@ -1,4 +1,4 @@
-package com.sleepkqq.sololeveling.ui.service.auth;
+package com.sleepkqq.sololeveling.ui.service;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
