+++ solo-leveling-ui-service/src/main/java/com/sleepkqq/sololeveling/ui/view/task/SwipedComponent.java
