+++ solo-leveling-ui-service/src/main/resources/META-INF/resources/icons/icon.png
