+++ solo-leveling-ui-service/src/main/resources/messages.properties
