+++ solo-leveling-ui-service/src/main/resources/messages_ru.properties
