+++ solo-leveling-ui-service/src/test/java/com/sleepkqq/sololeveling/ui/ApplicationTest.java
