+++ solo-leveling-ui-service/types.d.ts
